Read from PTY indefinitely
This uses the POSIX select() function, though it might not be ideal. We
might want to use polling instead.
Also, we specify a timeout of 0 to make select() non-blocking, which
kind of defeats its purpose. As soon as we can render anything in our
window and react to key events, we can start decoupling the reading from
the rendering. We have to find a way to share the "current screen" info
between these two processes.