Fix call to execvp
The manpage for exec(3) says that the call to execvp expects a
NULL-terminated array of CStrings as its second argument. Its first
element is supposed to be the (redundant) name/path of the binary that
is to be executed.

Most shells I tried (dash, bash, fish) worked as expected nevertheless,
but when I tried out zsh, it gave me a weird EIO. Doing this it started
working again.